Removed empty files, added request schema inside request.py
