+++ server/app/schemas/requests.py
@@ -35,6 +35,68 @@ class BaseRequest(BaseModel):
     filter_value: str = Field("", description="Search based on value")
 
 
+class PipelineNameRequest(BaseModel):
+    pipeline_name: str = Field(..., min_length=1, description="Name of the pipeline")
+
+
+class StageNameRequest(BaseModel):
+    stage_name: str = Field(..., min_length=1, description="Name of the stage")
+
+
+class ExecutionIdsRequest(BaseModel):
+    exe_ids: list[int] = Field(..., min_items=1, description="List of execution identifiers")
+
+
+class ArtifactIdsRequest(BaseModel):
+    artifact_ids: list[int] = Field(..., min_items=1, description="List of artifact identifiers")
+
+
+class ArtifactNameRequest(BaseModel):
+    artifact_name: str = Field(..., min_length=1, description="Name of the artifact")
+
+
+class ArtifactNameWithPipelineRequest(BaseModel):
+    artifact_name: str = Field(..., min_length=1, description="Name of the artifact")
+    pipeline_id: Optional[int] = Field(None, description="Optional pipeline identifier")
+
+
+class ArtifactIdRequest(BaseModel):
+    artifact_id: int = Field(..., description="Artifact identifier")
+
+
+class ExecutionIdRequest(BaseModel):
+    execution_id: int = Field(..., description="Execution identifier")
+
+
+class ExecutionIdsWithPipelineRequest(BaseModel):
+    """Execution id list request used by parent execution lookup endpoints."""
+    execution_id: list[int] = Field(..., min_items=1, description="List of execution identifiers")
+    pipeline_id: Optional[int] = Field(None, description="Optional pipeline identifier")
+
+
+class ParentExecutionIdRequest(BaseModel):
+    execution_id: int = Field(..., description="Execution identifier")
+    pipeline_id: Optional[int] = Field(None, description="Optional pipeline identifier")
+
+
+class StageIdRequest(BaseModel):
+    stage_id: int = Field(..., description="Stage identifier")
+    execution_uuid: Optional[str] = Field(None, description="Optional execution UUID")
+
+
+class MetricsNameRequest(BaseModel):
+    metrics_name: str = Field(..., min_length=1, description="Name of the metrics artifact")
+
+
+class PipelineJsonRequest(BaseModel):
+    pipeline_name: str = Field(..., min_length=1, description="Name of the pipeline")
+    exec_uuid: Optional[str] = Field(None, description="Optional execution UUID")
+
+
+class LastSyncTimeRequest(BaseModel):
+    last_sync_time: int = Field(..., description="Last sync time in epoch milliseconds")
+
+
 # Query parameters for execution.
 # class ExecutionRequest(BaseRequest):
 #     sort_field: str = Field("Context_Type", description="Column to sort by (default: Context_Type)")
